refactor(editor): extract editor shell copy applier
