+++ pages/editor.html
@@ -108,6 +108,7 @@
     <script src="../js/editor/editor-data-loader-fallbacks.js?v=20260422-1"></script>
     <script src="../js/editor/editor-entry-fallbacks.js?v=20260422-1"></script>
     <script src="../js/editor/editor-shell-helpers.js?v=20260422-1"></script>
+    <script src="../js/editor/editor-shell-copy-applier.js?v=20260522-1276"></script>
 
     <script src="../js/editor.js?v=20260502-1"></script>
     <script src="../js/editor/editor-i18n-refresh.js?v=20260422-2"></script>
